requirements.txt
pytest==8.1.1
PyYAML==6.0.1
Requests==2.31.0
